chore(eslint): Configure linting rules
- Fail linting when warnings are reported
- Rename ESLint configuration from `.js` to `.ts`
- Ignore coverage and worker-configuration.d.ts
